+++ resources/views/app/validator-monitor.blade.php
@@ -7,14 +7,6 @@
             :subtitle="trans('pages.validator-monitor.subtitle')"
         />
 
-        <div class="px-6 pb-6 md:px-10 md:mx-auto md:max-w-7xl">
-            <livewire:validator-data-boxes />
-        </div>
-
-        <x-general.mobile-divider />
-
-        <div class="px-6 pt-6 pb-8 md:px-10 md:pt-0 md:mx-auto md:max-w-7xl">
-            <livewire:validators.monitor />
-        </div>
+        <livewire:validators.monitor />
     @endsection
 @endcomponent
